Used timestamp_t where appropriate in readers and subtitles

+++ src/filesystem_reader.cpp
@@ -106,8 +106,8 @@ std::unique_ptr<input_reader> make_filesystem_reader(
 	std::string &input_file,
 	double fps,
 	std::string &audio_arg,
-	double from_index,	//	bad name, is a timestamp
-	double to_index
+	size_t from_index,	//	bad name, is a timestamp
+	size_t to_index
 )
 {
 	return std::make_unique<filesystem_reader>(input_file, fps, audio_arg, from_index, to_index);
