+++ src/filesystem_reader.hpp
@@ -6,6 +6,6 @@ std::unique_ptr<input_reader> make_filesystem_reader(
 	std::string &input_file,
 	double fps,
 	std::string &audio_arg,
-	double from_index,	//	bad name, is a timestamp
-	double to_index
+	size_t from_index,
+	size_t to_index
 );
